+++ src/test/java/mars/AppTest.java
@@ -1,6 +1,7 @@
 package mars;
 
 import com.vividsolutions.jts.geom.CoordinateList;
+import jdk.nashorn.internal.ir.annotations.Ignore;
 import junit.framework.Test;
 import junit.framework.TestCase;
 import junit.framework.TestSuite;
@@ -56,25 +57,6 @@ public class AppTest extends TestCase {
     }
 
     //@Test
-    public void testTIFFGetValueValid() throws Exception {
-        GeoTIFF newMap = new GeoTIFF();
-        newMap.initMap("src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff");
-        Double result = newMap.getValue(200,200);
-        assertNotSame(Double.toString(result),0,result);
-    }
-
-    //@Test
-    public void testTIFFGetValueBad() throws Exception {
-        GeoTIFF newMap = new GeoTIFF();
-        newMap.initMap("src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff");
-        try {
-            Double result = newMap.getValue(-1,-1);
-        } catch (Exception expectedException) {
-            assertEquals("Bad getValue",expectedException.getMessage());
-        }
-    }
-
-    //@Test
     public void testRoverGetSlopeZero() throws Exception {
         Coordinate starts = new Coordinate(10,10);
         Coordinate ends = new Coordinate(20, 20);
@@ -92,28 +74,6 @@ public class AppTest extends TestCase {
         Coordinate point1 = new Coordinate(20, 20);
         Coordinate point2 = new Coordinate(21, 20);
         assertTrue(newRover.canTraverse(point1,point2));
-    }
-
-    public void testGeotiffMaxHeight() throws Exception{
-        GeoTIFF newMap = new GeoTIFF();
-        newMap.initMap("src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff");
-        Double result = newMap.getMaxValue();
-        assertEquals(result, result);
-    }
-
-    public void testGeotiffMinHeight() throws Exception{
-        GeoTIFF newMap = new GeoTIFF();
-        newMap.initMap("src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff");
-        Double result = newMap.getMinValue();
-        assertEquals(result, result);
-    }
-
-    public void testGeotiffMinlessthanMax() throws Exception{
-        GeoTIFF newMap = new GeoTIFF();
-        newMap.initMap("src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff");
-        Double minresult = newMap.getMinValue();
-        Double maxresult = newMap.getMaxValue();
-        assertTrue(minresult < maxresult);
     }
 
 //     public void testGeotiffGetElevationsInArea() throws Exception {
@@ -140,91 +100,6 @@ public class AppTest extends TestCase {
 //         } catch (Exception e) { fail(); }
 //     }
 
-    public void testCLIcheckMapFailsWithoutRealFile() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertFalse(ti.checkMap(new Scanner("not a real file")));
-    }
-
-    public void testCLIcheckMapPassesWithRealFile() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertTrue(ti.checkMap(new Scanner("src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff")));
-    }
-
-    public void testCLIcheckSlopeFailsWithNonNumber() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertFalse(ti.checkSlope(new Scanner("letters")));
-    }
-
-    public void testCLIcheckSlopePassesWithNumber0() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertTrue(ti.checkSlope(new Scanner("0")));
-    }
-
-    public void testCLIcheckSlopeFailsWithNumberBelow0() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertFalse(ti.checkSlope(new Scanner("-19")));
-    }
-
-    public void testCLIcheckSlopePassesWithIntBetween0and90() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertTrue(ti.checkSlope(new Scanner("19")));
-    }
-
-    public void testCLIcheckSlopePassesWithDoubleBetween0and90() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertTrue(ti.checkSlope(new Scanner("19.1")));
-    }
-
-    public void testCLIcheckSlopePassesWithNumber90() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertTrue(ti.checkSlope(new Scanner("90")));
-    }
-
-    public void testCLIcheckSlopePassesWithNumberAbove90() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertFalse(ti.checkSlope(new Scanner("112")));
-    }
-
-    public void testCLIcheckSlopePassesIgnoringExtraneousData() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertTrue(ti.checkSlope(new Scanner("19.1 extraneous data")));
-    }
-
-    public void testCLIcheckStartCoordsFailsWithNonNumbers() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertFalse(ti.checkStartCoords(new Scanner("non-number non-number")));
-    }
-
-    public void testCLIcheckStartCoordsFailsWithNonIntegers() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertFalse(ti.checkStartCoords(new Scanner("19.1 19.1")));
-    }
-
-    public void testCLIcheckStartCoordsPassesWithTwoInts() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertTrue(ti.checkStartCoords(new Scanner("19 19")));
-    }
-
-    public void testCLIcheckEndCoordsFailsWithNonNumbers() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertFalse(ti.checkEndCoords(new Scanner("non-number non-number")));
-    }
-
-    public void testCLIcheckEndCoordsFailsWithNonIntegers() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertFalse(ti.checkEndCoords(new Scanner("19.1 19.1")));
-    }
-
-    public void testCLIcheckEndCoordsPassesWithTwoInts() throws Exception{
-        TerminalInterface ti = new TerminalInterface();
-        assertTrue(ti.checkEndCoords(new Scanner("19 19")));
-    }
-
-    public void testOutFileOutputConstuctorPasses() throws Exception{
-        CoordinateList coords = new CoordinateList();
-        FileOutput fo = new FileOutput(coords);
-    }
-
     /*
     public void testOutFileOutputWritesToFile() throws Exception{
         CoordinateList coords = new CoordinateList();
@@ -235,7 +110,6 @@ public class AppTest extends TestCase {
         fo.writeToOutput();
     }
     */
-
 
     public void testGreedyCoordinateInstantiate() throws Exception{
         Coordinate testCoordinate = new Coordinate(20,20);
@@ -249,279 +123,6 @@ public class AppTest extends TestCase {
         assertEquals(45,neighbors.get(7).getDirection());
     }
 
-    public void testGreedyAlgorithmUnlimitedFlatCase() throws Exception{
-        Coordinate startCoord = new Coordinate(10,10);
-        Coordinate endCoord = new Coordinate(10,20);
-        String mapPath = "src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff";
-        MarsRover rover = new MarsRover(1,startCoord,endCoord,mapPath);
-        Algorithm algorithm = new AlgorithmGreedy(rover,"unlimited");
-        try {
-            algorithm.findPath();
-        } catch (Exception expectedException) {
-            assertTrue("Failed to find a route it should have",false);
-        }
-    }
 
-    public void testGreedyAlgorithmLimitedFlatCase() throws Exception{
-        Coordinate startCoord = new Coordinate(10,10);
-        Coordinate endCoord = new Coordinate(10,20);
-        String mapPath = "src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff";
-        MarsRover rover = new MarsRover(1,startCoord,endCoord,mapPath);
-        Algorithm algorithm = new AlgorithmGreedy(rover,"limited");
-        try {
-            algorithm.findPath();
-        } catch (Exception expectedException) {
-            assertTrue("Failed to find a route it should have",false);
-        }
-    }
-
-    public void testGreedyAlgorithmComparedFlatCase() throws Exception{
-        Coordinate startCoord = new Coordinate(10,10);
-        Coordinate endCoord = new Coordinate(10,20);
-        String mapPath = "src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff";
-        MarsRover rover = new MarsRover(1,startCoord,endCoord,mapPath);
-        Algorithm algorithmUnlimited = new AlgorithmGreedy(rover,"unlimited");
-        Algorithm algorithmLimited = new AlgorithmGreedy(rover,"limited");
-        try {
-            algorithmUnlimited.findPath();
-            algorithmLimited.findPath();
-        } catch (Exception expectedException) {
-            assertTrue("Failed to find a route it should have",false);
-        }
-        int unlimitedLength = (algorithmUnlimited.getPath()).size();
-        int limitedLength = (algorithmLimited.getPath()).size();
-        boolean check = unlimitedLength == limitedLength;
-        assertTrue("Limited and unlimited routes are not the same",check);
-    }
-
-    /* Requires viking map to run, which makes Travis sad
-    public void testGreedyAlgorithmComparedComplexCase() throws Exception{
-        Coordinate startCoord = new Coordinate(7568,1507);
-        Coordinate endCoord = new Coordinate(7436,3927);
-        String mapPath = "src/main/resources/Phobos_Viking_Mosaic_40ppd_DLRcontrol.tif";
-        MarsRover rover = new MarsRover(30,startCoord,endCoord,mapPath);
-        Algorithm algorithmUnlimited = new AlgorithmGreedy(rover,"unlimited");
-        Algorithm algorithmLimited = new AlgorithmGreedy(rover,"limited");
-        try {
-            algorithmUnlimited.findPath();
-            algorithmLimited.findPath();
-        } catch (Exception expectedException) {
-            assertTrue("Failed to find a route it should have",false);
-        }
-        int unlimitedLength = (algorithmUnlimited.getPath()).size();
-        int limitedLength = (algorithmLimited.getPath()).size();
-        boolean check = unlimitedLength == limitedLength;
-        assertTrue("Limited and unlimited routes are not the same",check);
-    }
-    */
-
-    public void testGreedyAlgorithmUnlimitedSlopedCase() throws Exception{
-        Coordinate startCoord = new Coordinate(538,191);
-        Coordinate endCoord = new Coordinate(208,210);
-        String mapPath = "src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff";
-        MarsRover rover = new MarsRover(45,startCoord,endCoord,mapPath);
-        Algorithm algorithm = new AlgorithmGreedy(rover,"unlimited");
-        try {
-            algorithm.findPath();
-        } catch (Exception expectedException) {
-            assertTrue("Failed to find a route it should have",false);
-        }
-    }
-
-    public void testAstarAlgorithmLimitedSlopedCase() throws Exception{
-        Coordinate startCoord = new Coordinate(538,191);
-        Coordinate endCoord = new Coordinate(208,210);
-        String mapPath = "src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff";
-        MarsRover rover = new MarsRover(45,startCoord,endCoord,mapPath,3);
-        Algorithm algorithm = new AlgorithmLimitedScopeAStar(rover);
-        try {
-            algorithm.findPath();
-        } catch (Exception expectedException) {
-            assertTrue("Failed to find a route it should have",false);
-        }
-    }
-
-    public void testAstarAlgorithmComparedSlopedCase() throws Exception{
-        Coordinate startCoord = new Coordinate(538,191);
-        Coordinate endCoord = new Coordinate(208,210);
-        String mapPath = "src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff";
-        MarsRover rover = new MarsRover(45,startCoord,endCoord,mapPath,3);
-        Algorithm limitedAlgorithm = new AlgorithmLimitedScopeAStar(rover);
-        Algorithm unlimitedAlgorithm = new AlgorithmUnlimitedScopeRecursive(rover);
-        try {
-            limitedAlgorithm.findPath();
-            unlimitedAlgorithm.findPath();
-        } catch (Exception expectedException) {
-            assertTrue("Failed to find a route it should have",false);
-        }
-        int limitedLength = (limitedAlgorithm.getPath()).size();
-        int unlimitedLength = (unlimitedAlgorithm.getPath()).size();
-        boolean check = unlimitedLength < limitedLength;
-        assertTrue("Limited algorithm performed better",check);
-    }
-
-    public void testAstarAlgorithmLimitedFlatCase() throws Exception{
-        Coordinate startCoord = new Coordinate(10,10);
-        Coordinate endCoord = new Coordinate(10,20);
-        String mapPath = "src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff";
-        MarsRover rover = new MarsRover(1,startCoord,endCoord,mapPath,3);
-        Algorithm algorithm = new AlgorithmLimitedScopeAStar(rover);
-        try {
-            algorithm.findPath();
-        } catch (Exception expectedException) {
-            assertTrue("Failed to find a route it should have",false);
-        }
-    }
-
-    public void testAstarAlgorithmUnliimitedFlatCase() throws Exception{
-        Coordinate startCoord = new Coordinate(10,10);
-        Coordinate endCoord = new Coordinate(10,20);
-        String mapPath = "src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff";
-        MarsRover rover = new MarsRover(1,startCoord,endCoord,mapPath);
-        Algorithm algorithm = new AlgorithmUnlimitedScopeRecursive(rover);
-        try {
-            algorithm.findPath();
-        } catch (Exception expectedException) {
-            assertTrue("Failed to find a route it should have",false);
-        }
-    }
-
-    /*
-    public void testGreedyAlgorithmUnlimitedRoverSlopeMatters() throws Exception{
-        Coordinate startCoord = new Coordinate(7568,1507);
-        Coordinate endCoord = new Coordinate(7467,2625);
-        String mapPath = "src/main/resources/Phobos_Viking_Mosaic_40ppd_DLRcontrol.tif";
-        MarsRover strongRover = new MarsRover(90,startCoord,endCoord,mapPath);
-        MarsRover mediumRover = new MarsRover(40,startCoord,endCoord,mapPath);
-        MarsRover weakRover = new MarsRover(30,startCoord,endCoord,mapPath);
-        Algorithm algorithmStrong = new AlgorithmGreedy(strongRover,"unlimited");
-        Algorithm algorithmMedium = new AlgorithmGreedy(mediumRover,"unlimited");
-        Algorithm algorithmWeak = new AlgorithmGreedy(weakRover,"unlimited");
-        try {
-            algorithmStrong.findPath();
-            algorithmMedium.findPath();
-            algorithmWeak.findPath();
-        } catch (Exception expectedException) {
-            assertTrue("Failed to find a route it should have",false);
-        }
-        int strongLength = (algorithmStrong.getPath()).size();
-        int mediumLength = (algorithmMedium.getPath()).size();
-        int weakLength = (algorithmWeak.getPath()).size();
-        boolean check = strongLength < mediumLength && mediumLength < weakLength;
-        assertTrue("Stronger rovers aren't performing better",check);
-    }
-    */
-
-    /*
-    public void testAStarAlgorithmLimited() throws Exception{
-        Coordinate startCoord = new Coordinate(7568,1507);
-        Coordinate endCoord = new Coordinate(7467,2625);
-        //Coordinate startCoord = new Coordinate(538,191);
-        //Coordinate endCoord = new Coordinate(208,210);
-        //String mapPath = "src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff";
-        String mapPath = "src/main/resources/Phobos_Viking_Mosaic_40ppd_DLRcontrol.tif";
-        MarsRover rover = new MarsRover(90 ,startCoord,endCoord,mapPath,20);
-        Algorithm algorithm = new AlgorithmLimitedScopeAStar(rover);
-        try {
-            algorithm.findPath();
-        } catch (Exception expectedException) {
-            assertTrue("Failed to find a route it should have",false);
-        }
-        boolean check = true;
-        assertTrue(check);
-    }
-    */
-
-    public void testGreedyAlgorithmUnlimitedNoBacktrack() throws Exception{
-        Coordinate startCoord = new Coordinate(538,191);
-        Coordinate endCoord = new Coordinate(208,210);
-        String mapPath = "src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff";
-        MarsRover rover = new MarsRover(45,startCoord,endCoord,mapPath);
-        Algorithm algorithm = new AlgorithmGreedy(rover,"unlimited");
-        try {
-            algorithm.findPath();
-        } catch (Exception expectedException) {
-            assertTrue("Failed to find a route it should have",false);
-        }
-        ArrayList<? extends Coordinate> path = algorithm.getPath();
-        Set<Coordinate> pathSet = new HashSet<Coordinate>(path);
-        boolean check = path.size() == pathSet.size();
-        assertTrue("Path has duplicates",check);
-    }
-
-    public void testGreedyAlgorithmUnlimitedValidRoute() throws Exception{
-        Coordinate startCoord = new Coordinate(538,191);
-        Coordinate endCoord = new Coordinate(208,210);
-        String mapPath = "src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff";
-        MarsRover rover = new MarsRover(45,startCoord,endCoord,mapPath);
-        Algorithm algorithm = new AlgorithmGreedy(rover,"unlimited");
-        try {
-            algorithm.findPath();
-        } catch (Exception expectedException) {
-            assertTrue("Failed to find a route it should have",false);
-        }
-        ArrayList<? extends Coordinate> path = algorithm.getPath();
-        Coordinate oldItem;
-        Coordinate item = startCoord;
-        for (Coordinate aPath : path) { //foreach coordinate in list
-            oldItem = item;
-            item = aPath;
-            int diff = abs(oldItem.getX() - item.getX()) + abs(oldItem.getY() - item.getY());
-            if (diff > 2) {
-                assertTrue("Path points too far from each other", false);
-            }
-        }
-    }
-
-    public void testGreedyAlgorithmUnlimitedFailure() throws Exception{
-        Coordinate startCoord = new Coordinate(0,0); //this is on an island in the map that the rover can't escape
-        Coordinate endCoord = new Coordinate(-1,-1);
-        String mapPath = "src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff";
-        MarsRover rover = new MarsRover(0,startCoord,endCoord,mapPath);
-        Algorithm algorithm = new AlgorithmGreedy(rover,"unlimited");
-        boolean failure = false;
-        try { //it's supposed to fail, so catch results in a successful test
-            algorithm.findPath();
-            failure = true;
-        } catch (Exception expectedException) {
-            assertTrue(true);
-        }
-        if(failure){assertTrue("Found a path it shouldn't have",false);}
-    }
-
-    /*
-    public void testGreedyAlgorithmLimitedComparison() throws Exception{
-        Coordinate startCoord = new Coordinate(400,131);
-        Coordinate endCoord = new Coordinate(450,154);
-        String mapPath = "src/test/resources/Phobos_ME_HRSC_DEM_Global_2ppd.tiff";
-        MarsRover rover = new MarsRover(25,startCoord,endCoord,mapPath);
-        Algorithm algorithmUnlimited = new AlgorithmGreedy(rover,"unlimited");
-        Algorithm algorithmLimited = new AlgorithmGreedy(rover,"limited");
-        try {
-            algorithmUnlimited.findPath();
-            algorithmLimited.findPath();
-        } catch (Exception expectedException) {
-            assertEquals("Failed to find a route it should have", false, true);
-        }
-        int unlimitedLength = (algorithmUnlimited.getPath()).size();
-        int limitedLength = (algorithmLimited.getPath()).size();
-        boolean check = unlimitedLength != limitedLength;
-        assertTrue("Limited and unlimited routes are the same",check);
-    }
-    */
-
-    /*
-    //@Test
-    public void testReadBigTIFF() throws Exception {
-        Route newRoute = new Route();
-        String[] testArgs = { "src/main/resources/Mars_MGS_MOLA_DEM_mosaic_global_463m.tif" }; //file should be in src/main/resources
-        newRoute.main(testArgs);
-        Double resulttop = newRoute.getValue(0.0,0.0);
-        Double resultbottom = newRoute.getValue(0.0,23000.0);
-        boolean testSuccess = false;
-        if(resultbottom > resulttop){testSuccess = true;}
-        assertTrue(Double.toString(resultbottom - resulttop),testSuccess);
-    }
-    */
 
 }
